long tables

+++ R/sql_demographic.r
@@ -56,7 +56,7 @@ sql.load_file <- function(path) {
 #' @export sql.newdb
 sql.newdb <- function(schema=NULL) {
     if (is.null(schema))
-        schema <- paste(find.package("ccdata"), "sql/create_table.sql", sep="/") 
+        schema <- paste(find.package("ccdata"), "inst/sql/create_table.sql", sep="/") 
 
     sql.init.postgres()
     cmd <- sql.load_file(schema)
@@ -66,13 +66,52 @@ sql.newdb <- function(schema=NULL) {
 #' Add demographic table to the postgreSQL table.
 #' @param record
 #' @export sql.add.demographic
-sql.add.demographic <- function(record) {
+sql.add.demographic <- function(record, update=FALSE) {
     sql.init.postgres()
-    sqldf('DROP TABLE IF EXISTS demographic;')
+    if (update)
+        sqldf('DROP TABLE IF EXISTS demographic;')
     drv <- dbDriver("PostgreSQL")
     con <- dbConnect(drv, host="localhost", user="sinan", dbname="postgres")
     demogt <- sql.demographic.table(record, dtype=TRUE)
     dbWriteTable(con, "demographic", demogt)
     dbDisconnect(con)
     invisible(demogt)
+}
+
+
+episode.long.table <- function(ep) {
+    eid <- paste(ep@site_id, ep@episode_id, sep="_")
+    episode_data <- list()
+    for(fname in names(ep@data)) {
+        field <- ep@data[[fname]]
+        nrows <- nrow(field)
+        if (!is.null(nrows)) {#2d data
+            d <- list()
+            stopifnot(!is.null(field$time))
+            stopifnot(!is.null(field$item2d))
+            d[["eid"]] <- rep(eid, nrows)
+            d[["field"]] <- rep(fname, nrows)
+            d[["time"]] <- field$time
+            d[["value"]] <- field$item2d
+            d[["meta"]] <- field$meta
+            if (is.null(d[["meta"]]))
+                d[["meta"]] <- rep(NA, nrows)
+            episode_data[[length(episode_data) + 1]] <- .simple.data.frame(d)
+        }
+    }
+    return(rbindlist(episode_data))
+}
+
+#'
+#' @export sql.add.episode
+sql.add.episode <- function(record, update=FALSE) {
+    sql.init.postgres()
+    tb <- "timedata"
+    if (update)
+        sqldf(paste('DROP TABLE IF EXISTS', tb))
+    drv <- dbDriver("PostgreSQL")
+    con <- dbConnect(drv, host="localhost", user="sinan", dbname="postgres")
+    for_each_episode(record, 
+        function(ep)
+            dbWriteTable(con, tb, episode.long.table(ep), append=T))
 }
